requirements.txt
requests
milvus-pymilvus
fastapi
uvicorn
jinja2
python-multipart
pyyaml
